Update acpi submodule

+++ kernel/kernel_x86_64/src/topo.rs
@@ -10,7 +10,7 @@ pub const BOOT_CPU_ID: CpuId = 0;
 
 pub struct Cpu {
     pub id: CpuId,
-    pub local_apic_id: u8,
+    pub local_apic_id: u32,
     pub per_cpu: Pin<Box<PerCpuImpl>>,
 }
